refactor: refactor model data type and logic code for use GORM's feature

+++ dtos/comment_dto.go
@@ -3,18 +3,18 @@ package dtos
 import "time"
 
 type CommentDTO struct {
-	ID        uint64        `json:"id"`
+	ID        uint          `json:"id"`
 	Author    SimpleUserDTO `json:"author"`
-	PostID    uint64        `json:"postId"`
+	PostID    uint          `json:"postId"`
 	Body      string        `json:"body"`
 	CreatedAt time.Time     `json:"createdAt"`
 	Reactions []ReactionDTO `json:"reactions"`
 }
 
 type SimpleCommentDTO struct {
-	ID        uint64        `json:"id"`
+	ID        uint          `json:"id"`
 	Author    SimpleUserDTO `json:"author"`
-	PostID    uint64        `json:"postId"`
+	PostID    uint          `json:"postId"`
 	Body      string        `json:"body"`
 	CreatedAt time.Time     `json:"createdAt"`
 }
