+++ models/content_reaction.go
@@ -4,12 +4,12 @@ import "gorm.io/gorm"
 
 type ContentReaction struct {
 	gorm.Model
-	Type      string  `gorm:"size:255;not null"`
-	PostID    *uint64 `gorm:""`
-	CommentID *uint64 `gorm:""`
-	MessageID *uint64 `gorm:""`
-	StoryID   *uint64 `gorm:""`
-	Count     int     `gorm:"default:0"`
+	Type      string `gorm:"size:255;not null"`
+	PostID    *uint  `gorm:""`
+	CommentID *uint  `gorm:""`
+	MessageID *uint  `gorm:""`
+	StoryID   *uint  `gorm:""`
+	Count     int    `gorm:"default:0"`
 }
 
 func (*ContentReaction) TableName() string {
